+++ Changelog.md
@@ -7,6 +7,12 @@ and this project follows the [Ragnarök Versioning Convention](https://github.co
 
 ## Wither Config Version 1.2 Changelog - 2024-06-24
 
+### Added
+
+- Configuration for:
+  - Invulnerable Healing
+  - Vulnerable Healing
+
 ### Changed
 
 - Updated dependencies: